First select pads for current user, then if it's anonymous user and it's preferred so return public knowledge pads for context. This change allows logged in users to have their own customized pads independently from value of a web site preference : "Force Anonymous Gadget " and still we can show or hide public pads for context for anonymous.
In case still we want to show default pads for logged in users we can use preferences which can contain knowledge pads templates. The templates can itself be customized which allows a KM or ERP5 UI site to show a set of default pads for any context which with a single click can be adjusted by user to his needs.